Separating Board into abstract class (More needs to be done in the abstract class)
