+++ src/Sudoku.java
@@ -53,7 +53,7 @@ public class Sudoku {
        		 			  { 0, 4, 0, 0, 5, 0, 0, 3, 6 },
        		 			  { 7, 0, 3, 0, 1, 8, 0, 0, 0 } };
 		
-		Board board = new Board(grid2);
+		BasicBoard board = new BasicBoard(grid2);
 		System.out.println(board.toString());
 		board.solve();
 		System.out.println(board.toString());
